Render buffer - a struct holding buffers that we will draw text from.
Will probably become obsolete once if we get to syntax highlighting.

Requirements for handling huge files with gigantic lines: must change
to be more integrated with the rope - either fns in the rope to
extract lines (make the rope aware of the render buffer) or use public
row api to extract small sub-ropes based on lines (make the render
buffer aware of the rope). Second is easiest.